refactor(upstream): keep everything in one package
