+++ pkg/upstream/load_test.go
@@ -1,4 +1,4 @@
-package loader
+package upstream
 
 import (
 	"io/ioutil"
@@ -28,7 +28,7 @@ upstreams:
 		assert.FailNow(t, "Error while writing to temp file", err)
 	}
 
-	loadedUpstreams, err := loadUpstreamsFromFile(tempFile)
+	loadedUpstreams, err := LoadFromFile(tempFile)
 	require.Nil(t, err, "Should load upstreams correctly")
 
 	require.Equal(t, 2, len(loadedUpstreams))
